Move interpolate methods to math utils
